Consolidate mongo-shell boundary into mongo-shell/ directory
Move shell/mongo-shell-adapter.js → mongo-shell/adapter.js and
cli/mongo-shell-launcher.js → mongo-shell/launcher.js, creating a single
mongo-shell/ package boundary for everything that concerns the Mongo shell.

Why mongo-shell/ is the right home for both files:
- adapter.js is the shell-script layer (assembled into variety.js by build.js);
  launcher.js is the Node-side process spawner (spawnSync for mongosh/mongo).
  Both files are specifically about the Mongo shell as a transport, so grouping
  them together makes the boundary self-documenting.
- Future interfaces (MCP, programmatic API, etc.) can import from mongo-shell/
  without reaching into cli/, which stays focused on the bin/variety CLI surface.

This is a pure rename — no logic changes. All import paths in cli/main.js,
test/cli/CliOptionsTest.js, build.js, .eslint.config.js,
.tsconfig.checkjs.json, and CONTRIBUTING.md are updated to match. variety.js
is regenerated (build output now references mongo-shell/adapter.js).
