+++ .eslint.config.js
@@ -79,17 +79,17 @@ module.exports = [
     },
   },
   {
-    files: ['.eslint.config.js', 'bin/variety', 'build.js', 'cli/**/*.js', 'test/**/*.js'],
+    files: ['.eslint.config.js', 'bin/variety', 'build.js', 'cli/**/*.js', 'mongo-shell/launcher.js', 'test/**/*.js'],
     ignores: ['test/fixtures/**/*.js'],
     rules: nodeModernizationRules,
   },
   {
-    files: ['variety.js', 'core/analyzer.js', 'shell/mongo-shell-adapter.js'],
+    files: ['variety.js', 'core/analyzer.js', 'mongo-shell/adapter.js'],
     // Rules safe for the ES6+ compatibility floor declared at the top of variety.js.
     // prefer-object-has-own is excluded because Object.hasOwn() (ES2022) is absent
     // from the legacy mongo shell; see the compatibility comment in variety.js.
     // variety.js is generated by build.js from the core/analyzer.js and
-    // shell/mongo-shell-adapter.js source files above; we lint both the sources
+    // mongo-shell/adapter.js source files above; we lint both the sources
     // and the committed build output so either entry point catches regressions.
     rules: {
       'no-throw-literal': 'error',
