+++ build.js
@@ -2,7 +2,7 @@
 'use strict';
 
 // Assembles variety.js by concatenating the formatter files, core/analyzer.js, and
-// shell/mongo-shell-adapter.js underneath a generated-file banner. The built variety.js
+// mongo-shell/adapter.js underneath a generated-file banner. The built variety.js
 // is committed to the repository so `mongosh variety.js` works from a fresh clone without
 // running a build step first; CI runs this script and fails if the
 // committed file drifts from its sources.
@@ -26,7 +26,7 @@ Released by James Cropcho, © 2012–2026, under the MIT License. */
 //
 // Assembled by build.js from:
 //   core/formatters/ascii.js, core/formatters/json.js,
-//   core/analyzer.js, shell/mongo-shell-adapter.js.
+//   core/analyzer.js, mongo-shell/adapter.js.
 // To change behavior, edit those source files and run \`npm run build\`. The
 // build output is committed so \`mongosh variety.js\` works from a fresh clone
 // without a build step; CI verifies the committed file matches its sources.
@@ -55,7 +55,7 @@ Released by James Cropcho, © 2012–2026, under the MIT License. */
 //      as explicit parameters. The section hands a bundle of functions to
 //      the interface section via \`shellContext.__varietyImpl\`.
 //
-//   3. INTERFACE SECTION (shell/mongo-shell-adapter.js) — everything that touches
+//   3. INTERFACE SECTION (mongo-shell/adapter.js) — everything that touches
 //      shell globals: reading input (\`collection\`, \`plugins\`, \`__quiet\`,
 //      \`slaveOk\`, etc.), the config-echo logging, plugin loading via
 //      \`load()\`, input validation, and constructing the dependency bag
@@ -70,7 +70,7 @@ const root = __dirname;
 const fmtAscii = fs.readFileSync(path.join(root, 'core', 'formatters', 'ascii.js'), 'utf8');
 const fmtJson  = fs.readFileSync(path.join(root, 'core', 'formatters', 'json.js'), 'utf8');
 const impl     = fs.readFileSync(path.join(root, 'core', 'analyzer.js'), 'utf8');
-const iface    = fs.readFileSync(path.join(root, 'shell', 'mongo-shell-adapter.js'), 'utf8');
+const iface    = fs.readFileSync(path.join(root, 'mongo-shell', 'adapter.js'), 'utf8');
 
 // HEADER ends with a single \n; each source file ends with a single \n. We
 // want two blank lines (three \n total) between each region in the output.
@@ -83,7 +83,7 @@ if (args.includes('--check')) {
   const existing = fs.readFileSync(outPath, 'utf8');
   if (existing !== output) {
     process.stderr.write(
-      'variety.js is out of date relative to its sources (core/formatters/, core/analyzer.js, shell/mongo-shell-adapter.js).\n' +
+      'variety.js is out of date relative to its sources (core/formatters/, core/analyzer.js, mongo-shell/adapter.js).\n' +
       'Run `npm run build` and commit the updated variety.js.\n'
     );
     process.exit(1);
