+++ mongo-shell/adapter.js
