+++ mongo-shell/launcher.js
